BACKEND/requirements.txt
django
djangorestframework
django-cors-headers
numpy
tifffile
